+++ lnx-page-store/src/file/page_op_log/tests/writer_associated_data.rs
@@ -14,7 +14,7 @@ use crate::layout::log::{LogEntry, LogOp};
 async fn test_single_block_correct_associated_data_tagging(
     #[values(0, 4096)] log_offset: u64,
 ) {
-    const FILE_ID: u64 = 1;
+    const FILE_ID: u32 = 1;
 
     let _ = tracing_subscriber::fmt::try_init();
 
@@ -55,7 +55,7 @@ async fn test_single_block_correct_associated_data_tagging(
 async fn test_multi_block_correct_associated_data_tagging(
     #[values(0, 4096)] log_offset: u64,
 ) {
-    const FILE_ID: u64 = 1;
+    const FILE_ID: u32 = 1;
 
     let _ = tracing_subscriber::fmt::try_init();
